Adjust layout using view pager
