+++ app/src/main/java/com/portto/valuedapp/evm/EvmChain.kt
@@ -4,6 +4,7 @@ import com.portto.valuedapp.Config
 
 enum class EvmChain(
     val title: String,
+    val symbol: String,
     val mainnetContractAddress: String,
     val testnetContractAddress: String,
     val mainnetRpcUrl: String,
@@ -14,6 +15,7 @@ enum class EvmChain(
 
     ETHEREUM(
         title = "Ethereum",
+        symbol = "ETH",
         mainnetContractAddress = "0x806243c7368a90D957592B55875eF4C3353C5bEa",
         testnetContractAddress = "0x58F385777aa6699b81f741Dd0d5B272A34C1c774",
         mainnetRpcUrl = "https://mainnet.infura.io/v3/${Config.INFURA_ID}",
@@ -23,6 +25,7 @@ enum class EvmChain(
     ),
     BNB_CHAIN(
         title = "BNB Chain",
+        symbol = "BNB",
         mainnetContractAddress = "0x806243c7368a90D957592B55875eF4C3353C5bEa",
         testnetContractAddress = "0xfde90c9Bc193F520d119302a2dB8520D3A4408c8",
         mainnetRpcUrl = "https://bsc-dataseed.binance.org",
@@ -32,6 +35,7 @@ enum class EvmChain(
     ),
     POLYGON(
         title = "Polygon",
+        symbol = "MATIC",
         mainnetContractAddress = "0x806243c7368a90D957592B55875eF4C3353C5bEa",
         testnetContractAddress = "0xfde90c9Bc193F520d119302a2dB8520D3A4408c8",
         mainnetRpcUrl = "https://polygon-mainnet.infura.io/v3/${Config.INFURA_ID}",
@@ -41,11 +45,12 @@ enum class EvmChain(
     ),
     AVALANCHE(
         title = "Avalanche",
+        symbol = "AVAX",
         mainnetContractAddress = "0x806243c7368a90D957592B55875eF4C3353C5bEa",
         testnetContractAddress = "0xfde90c9Bc193F520d119302a2dB8520D3A4408c8",
         mainnetRpcUrl = "https://api.avax.network/ext/bc/C/rpc",
         testnetRpcUrl = "https://api.avax-test.network/ext/bc/C/rpc",
         mainnetExplorerDomain = "avascan.info",
-        testnetExplorerDomain = "testnet.avascan.info"
+        testnetExplorerDomain = "testnet.snowtrace.io"
     )
 }
