+++ app/src/main/java/com/portto/valuedapp/evm/EvmPagerAdapter.kt
@@ -1,0 +1,17 @@
+package com.portto.valuedapp.evm
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.viewpager2.adapter.FragmentStateAdapter
+
+class EvmPagerAdapter(activity: FragmentActivity) : FragmentStateAdapter(activity) {
+
+    override fun getItemCount(): Int = 2
+
+    override fun createFragment(position: Int): Fragment {
+        return when (position) {
+            0 -> EvmSignMessageFragment()
+            else -> EvmSendTransactionFragment()
+        }
+    }
+}
