+++ app/src/main/res/drawable/bg_tab.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/bloctoBlue" />
+
+    <corners android:radius="20dp" />
+
+</shape>
