+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,6 @@
     <string name="title_get_value">Value: </string>
     <string name="title_partial_sign">Create Account and Set Value (PartialSign)</string>
     <string name="title_donate">Donate</string>
+    <string name="title_sign_message">Sign Message</string>
+    <string name="title_send_transaction">Send Transaction</string>
 </resources>
