+++ build.gradle
@@ -5,8 +5,8 @@ buildscript {
     }
 }
 plugins {
-    id 'com.android.application' version '7.1.3' apply false
-    id 'com.android.library' version '7.1.3' apply false
+    id 'com.android.application' version '7.2.0' apply false
+    id 'com.android.library' version '7.2.0' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.21' apply false
     id 'org.jetbrains.kotlin.plugin.serialization' version '1.6.21'
 }
