+++ dependencies.gradle
@@ -7,14 +7,19 @@ ext.versions = [
         androidx_core       : '1.7.0',
         androidx_appcompat  : '1.4.1',
         androidx_runtime    : '2.4.1',
+        androidx_viewmodel  : '2.4.1',
         androidx_browser    : '1.4.0',
+        androidx_activity   : '1.4.0',
+        androidx_fragment   : '1.4.1',
         coroutine           : '1.6.1',
-        material            : '1.5.0',
+        material            : '1.6.0',
         constraintlayout    : '2.1.3',
         serialization       : '1.3.2',
         okhttp              : '4.9.3',
         kbase58             : '0.4',
         solana_web3         : '0.1.3',
+        web3j               : '4.8.7-android',
+        progressbutton      : '2.1.0',
 
         junit               : '4.13.2',
         mockk               : '1.12.3',
